Improve federation api get_keys
