+++ crates/ruma-client-api/src/r0/keys/upload_keys.rs
@@ -4,10 +4,8 @@ use std::collections::BTreeMap;
 
 use js_int::UInt;
 use ruma_api::ruma_api;
-use ruma_common::encryption::DeviceKeys;
+use ruma_common::encryption::{DeviceKeys, OneTimeKey};
 use ruma_identifiers::{DeviceKeyAlgorithm, DeviceKeyId};
-
-use super::OneTimeKey;
 
 ruma_api! {
     metadata: {
